move files to follow the Atomic Design system
