+++ src/components/atoms/main.tsx
@@ -1,0 +1,11 @@
+type MainProps = {
+  children: React.ReactNode;
+};
+
+import styles from "./main.module.scss";
+
+export default function Main(props: MainProps) {
+  const { children } = props;
+
+  return <main className={styles.main}>{children}</main>;
+}
